Move includes to new splash subdir
 - move include files
 - create splash dir in build dir
 - adjust internal includes
