+++ src/include/splash/CollectionType.hpp
