+++ src/include/splash/basetypes/basetypes_compound.hpp
@@ -19,46 +19,52 @@
  * If not, see <http://www.gnu.org/licenses/>. 
  */
 
-#ifndef BASETYPES_ARRAY_HPP
-#define	BASETYPES_ARRAY_HPP
+#ifndef BASETYPES_COMPOUND_HPP
+#define	BASETYPES_COMPOUND_HPP
 
 #include <stdint.h>
 
-#include "CollectionType.hpp"
+#include "splash/CollectionType.hpp"
 
 namespace splash
 {
-    
-#define TYPE_ARRAY(_name, _h5_type, _real_type, _size)                         \
-    class ColType##_name##Array : public CollectionType                        \
+
+    static const char* COMPOUND_ELEMENTS[] = {"x", "y", "z", "u", "v", "w"};
+
+#define TYPE_COMPOUND(_name, _h5_type, _real_type, _size)                      \
+    class ColType##_name : public CollectionType                               \
     {                                                                          \
     public:                                                                    \
                                                                                \
-        ColType##_name##Array()                                                \
+        ColType##_name()                                                       \
         {                                                                      \
-            const hsize_t dim[] = {_size};                                     \
-            this->type = H5Tarray_create(_h5_type, 1, dim);                    \
+            this->type = H5Tcreate(H5T_COMPOUND, sizeof (_real_type) * _size); \
+            for (size_t i = 0; i < _size; ++i) {                               \
+                H5Tinsert(this->type, COMPOUND_ELEMENTS[i],                    \
+                    i * sizeof(_real_type), _h5_type);                         \
+            }                                                                  \
         }                                                                      \
                                                                                \
-        ~ColType##_name##Array()                                               \
+        ~ColType##_name()                                                      \
         { H5Tclose(this->type); }                                              \
                                                                                \
         size_t getSize() const                                                 \
-        { return sizeof (_real_type) * _size; }                                \
+        { return _size * sizeof (_real_type); }                                \
     };
-    
-TYPE_ARRAY(Float2, H5T_NATIVE_FLOAT, float, 2);
-TYPE_ARRAY(Float3, H5T_NATIVE_FLOAT, float, 3);
-TYPE_ARRAY(Float4, H5T_NATIVE_FLOAT, float, 4);
 
-TYPE_ARRAY(Double2, H5T_NATIVE_DOUBLE, double, 2);
-TYPE_ARRAY(Double3, H5T_NATIVE_DOUBLE, double, 3);
-TYPE_ARRAY(Double4, H5T_NATIVE_DOUBLE, double, 4);
+TYPE_COMPOUND(Float2, H5T_NATIVE_FLOAT, float, 2);
+TYPE_COMPOUND(Float3, H5T_NATIVE_FLOAT, float, 3);
+TYPE_COMPOUND(Float4, H5T_NATIVE_FLOAT, float, 4);
 
-TYPE_ARRAY(Int2, H5T_NATIVE_INT, int, 2);
-TYPE_ARRAY(Int3, H5T_NATIVE_INT, int, 3);
-TYPE_ARRAY(Int4, H5T_NATIVE_INT, int, 4);
+TYPE_COMPOUND(Double2, H5T_NATIVE_DOUBLE, double, 2);
+TYPE_COMPOUND(Double3, H5T_NATIVE_DOUBLE, double, 3);
+TYPE_COMPOUND(Double4, H5T_NATIVE_DOUBLE, double, 4);
+
+TYPE_COMPOUND(Int2, H5T_NATIVE_INT, int, 2);
+TYPE_COMPOUND(Int3, H5T_NATIVE_INT, int, 3);
+TYPE_COMPOUND(Int4, H5T_NATIVE_INT, int, 4);
+
 }
 
-#endif	/* BASETYPES_ARRAY_HPP */
+#endif	/* BASETYPES_COMPOUND_HPP */
