+++ src/include/splash/core/DCGroup.hpp
@@ -29,9 +29,9 @@
 #include <vector>
 #include <hdf5.h>
 
-#include "DataCollector.hpp"
-#include "DCException.hpp"
-#include "core/HandleMgr.hpp"
+#include "splash/DataCollector.hpp"
+#include "splash/DCException.hpp"
+#include "splash/core/HandleMgr.hpp"
 
 namespace splash
 {
