+++ src/include/splash/core/SDCHelper.hpp
@@ -24,11 +24,11 @@
 #ifndef SDCHELPER_H
 #define	SDCHELPER_H
 
-#include "basetypes/ColTypeDim.hpp"
-#include "sdc_defines.hpp"
-#include "core/DCAttribute.hpp"
-#include "core/logging.hpp"
-#include "DCException.hpp"
+#include "splash/basetypes/ColTypeDim.hpp"
+#include "splash/sdc_defines.hpp"
+#include "splash/core/DCAttribute.hpp"
+#include "splash/core/logging.hpp"
+#include "splash/DCException.hpp"
 
 namespace splash
 {
