+++ src/include/splash/core/logging.hpp
