+++ src/include/splash/version.hpp
