+++ tests/include/DomainsTest.h
@@ -26,7 +26,7 @@
 
 #include <cppunit/extensions/HelperMacros.h>
 
-#include "splash.h"
+#include "splash/splash.h"
 
 using namespace splash;
