+++ tests/include/Parallel_AttributesTest.h
@@ -25,7 +25,7 @@
 
 #include <cppunit/extensions/HelperMacros.h>
 
-#include "splash.h"
+#include "splash/splash.h"
 
 using namespace splash;
